+++ test/tensor_indexing_test.rb
@@ -14,14 +14,29 @@ class TensorIndexingTest < Minitest::Test
     assert_equal [0, 1], x[0, 0...-1].to_a
     assert_equal [[0, 1, 2]], x[0...-1].to_a
     assert_equal [], x[false].to_a
-    if RUBY_VERSION.to_f > 2.6
-      assert_equal [1, 2], x[0, eval("1..")].to_a
-      assert_equal [1], x[0, eval("(1...)")].to_a
-    end
-    if RUBY_VERSION.to_f > 2.7
-      assert_equal [0, 1], x[0, eval("..2")].to_a
-      assert_equal [0, 1, 2], x[0, eval("..")].to_a
-    end
+  end
+
+  def test_getter_endless
+    skip if RUBY_VERSION.to_f < 2.6
+
+    x = Torch.tensor([[0, 1, 2], [3, 4, 5]])
+    assert_equal [1, 2], x[0, eval("1..")].to_a
+    assert_equal [1], x[0, eval("(1...)")].to_a
+    assert_equal [2], x[0, eval("-1..")].to_a
+    assert_equal [], x[0, eval("(-1...)")].to_a
+    assert_equal [1, 2], x[0, eval("-2..")].to_a
+    assert_equal [1], x[0, eval("(-2...)")].to_a
+  end
+
+  def test_getter_beginless
+    skip if RUBY_VERSION.to_f < 2.7
+
+    x = Torch.tensor([[0, 1, 2], [3, 4, 5]])
+    assert_equal [0, 1], x[0, eval("..1")].to_a
+    assert_equal [0], x[0, eval("...1")].to_a
+    assert_equal [0, 1, 2], x[0, eval("..-1")].to_a
+    assert_equal [0, 1], x[0, eval("..-2")].to_a
+    assert_equal [0], x[0, eval("...-2")].to_a
   end
 
   def test_unsupported_type